+++ code/LIMA_ENGINE/src/SupernaturalForces.cuh
@@ -81,7 +81,7 @@ namespace SupernaturalForces {
 	}
 
 	// Overwrites the force that is given as an argument to the function
-	__global__ void ApplyHorizontalSqueeze(SimulationDevice* simDev, int step) {
+	__global__ void ApplyHorizontalSqueeze(SimulationDevice* simDev, int64_t step) {
 		__shared__ Float3 avg_abspos_nm;
 		__shared__ float avg_force_z;
 		__shared__ Float3 relPosNm[MAX_COMPOUND_PARTICLES];
@@ -166,7 +166,7 @@ namespace SupernaturalForces {
 		return force;
 	}
 
-	__global__ void BoxEdgeForceCompounds(SimulationDevice* simDev, int step) {
+	__global__ void BoxEdgeForceCompounds(SimulationDevice* simDev, int64_t step) {
 		__shared__ Float3 origo;
 
 		const auto const coords = CompoundcoordsCircularQueueUtils::getCoordarrayRef(simDev->boxState->compoundcoordsCircularQueue, step, blockIdx.x);
@@ -179,7 +179,7 @@ namespace SupernaturalForces {
 			simDev->boxState->compounds[blockIdx.x].forces_interim[threadIdx.x] += BoxEdgeForce(positionNM) * massFactor;
 	}
 
-	__global__ void BoxEdgeForceSolvents(SimulationDevice* simDev, int step) {
+	__global__ void BoxEdgeForceSolvents(SimulationDevice* simDev, int64_t step) {
 		// TODONOW Implement
 		// 
 		// 
@@ -202,7 +202,7 @@ namespace SupernaturalForces {
 
 
 
-	void SnfHandler(Simulation* simulation, SimulationDevice* simDev, int step) {
+	void SnfHandler(Simulation* simulation, SimulationDevice* simDev, int64_t step) {
 		cudaDeviceSynchronize();
